Add multiple pages and tableau
